* 'test' of https://github.com/projectClone5/projectClone5-BE:
  jwt token issue fixed
  s3 설정중 - 1단계
  로그인기능 붙히기
  s3 그레이몬 추가
  406 에러 수정 완료
  컨트롤러 주석문 제거
  댓글 수정 , 삭제 , 작성

# Conflicts:
#	spring_clone_coding/spring_projectClone/build.gradle
#	spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/controller/CommentController.java
#	spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/controller/PostController.java
#	spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/controller/UserController.java
#	spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/model/Comment.java
#	spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/service/CommentService.java

+++ spring_clone_coding/spring_projectClone/src/main/java/com/sparta/spring_projectclone/service/CommentService.java
@@ -14,52 +14,55 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class CommentService {
 
-
     private final PostRepository postRepository;
     private final CommentRepository commentRepository;
 
 
     // 댓글 작성
     @Transactional
-    public CommentResponseDto commentWrite(Long postId, CommentRequestDto commentRequestDto) {
+    public CommentResponseDto commentWrite(Long postId, CommentRequestDto commentRequestDto , String username) {
 
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
-        Comment comment = new Comment(commentRequestDto, post);
+        Comment comment = new Comment(commentRequestDto , post, username);
+        System.out.println("옴????");
         CommentResponseDto commentResponseDto = new CommentResponseDto(comment);
-        //Post에 리뷰포인트, 코멘트 수 추가
-        post.modifyAvgReviewPoint(commentRequestDto.getReviewPoint(), 1);
+            post.modifyAvgReviewPoint(commentRequestDto.getReviewPoint(), 1);
         commentRepository.save(comment);
-
         return commentResponseDto;
-
-
-//        Comment comment = Comment.commentCreateDto(commentRequestDto);
-//        commentRepository.save(comment , post);
-//        CommentResponseDto commentResponseDto = new CommentResponseDto(comment);
-//        return commentResponseDto;
     }
 
     //댓글 수정
     @Transactional
-    public CommentResponseDto commentUpdated(Long commentId, CommentRequestDto commentRequestDto) {
+    public CommentResponseDto commentUpdated(Long commentId, CommentRequestDto commentRequestDto, String username){
         Comment comment = commentRepository.findById(commentId).orElseThrow(
                 () -> new IllegalArgumentException("댓글이 존재하지 않습니다."));
-
-        //Post에 리뷰포인트 수정
-        comment.getPost().modifyAvgReviewPoint(commentRequestDto.getReviewPoint() - comment.getReviewPoint(), 0);
-
-        comment.update(commentRequestDto);
-        commentRepository.save(comment);
-        return new CommentResponseDto(comment);
+        String writer = comment.getUsername();
+        System.out.println("옴?");
+        // 본인이 작성한 글이 아닙니다
+        if(writer.equals(username)){
+            //Post에 리뷰포인트 수정
+            comment.getPost().modifyAvgReviewPoint(commentRequestDto.getReviewPoint() - comment.getReviewPoint(), 0);
+            comment.update(commentRequestDto);
+            CommentResponseDto CommentResponseDto = new CommentResponseDto(comment);
+            return CommentResponseDto;
+        } else {
+            throw new IllegalArgumentException("작성한 유저가 아닙니다.");
+        }
     }
 
     // 댓글 삭제
-    public void commentDelete(Long commentId) {
+    public void commentDelete(Long commentId, String username) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(
                 () -> new IllegalArgumentException("댓글이 존재하지 않습니다."));
-        //Post에 리뷰포인트 삭제
-        comment.getPost().modifyAvgReviewPoint(-comment.getReviewPoint(), -1);
-        commentRepository.delete(comment);
+        String writer = comment.getUsername();
+        // 본인이 작성한 글이 아닙니다
+        if(writer.equals(username)){
+            //Post에 리뷰포인트 삭제
+            comment.getPost().modifyAvgReviewPoint(-comment.getReviewPoint(), -1);
+            commentRepository.delete(comment);
+        } else {
+            throw new IllegalArgumentException("작성한 유저가 아닙니다.");
+        }
     }
 }
